+++ services/paymaster/src/config.rs
@@ -30,9 +30,6 @@ pub struct Domain {
 
 #[derive(Deserialize)]
 pub struct Config {
-    pub mnemonic_file: String,
-    pub solana_url: String,
-    pub listen_address: String,
     pub domains: Vec<Domain>,
     #[serde(default)]
     pub tolls: Vec<Tolls>,