+++ packages/sessions-sdk-ts/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @fogo/sessions-sdk
 
+## 0.0.16
+
+### Patch Changes
+
+- 80cc3c3: Fix bug with legacy VersionedTransaction using Address Lookup Tables by bumping @solana/kit
+
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [0eead65]
+  - @fogo/sessions-idls@0.0.6
+
 ## 0.0.14
 
 ### Patch Changes