+++ scripts/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @fogo/scripts
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [80cc3c3]
+  - @fogo/sessions-sdk@0.0.16
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [0eead65]
+  - @fogo/sessions-idls@0.0.6
+  - @fogo/sessions-sdk@0.0.15
+
 ## 0.0.6
 
 ### Patch Changes